add: items index

+++ internal/app/purchase/repository.go
@@ -28,7 +28,7 @@ type OrderResult struct {
 	EstimatedDeliveryTimeInMinutes int32
 }
 
-func (r *PurchaseRepository) CreateEstimateWithOrders(ctx context.Context, userLat, userLng, totalPrice float64, estimatedTime int32, orders []Order) (EstimateResult, error) {
+func (r *PurchaseRepository) CreateEstimateWithOrders(ctx context.Context, userLat, userLng, totalPrice float64, estimatedTime int, orders []Order) (EstimateResult, error) {
 	var result EstimateResult
 
 	// Use transaction for performance and consistency
@@ -45,7 +45,7 @@ func (r *PurchaseRepository) CreateEstimateWithOrders(ctx context.Context, userL
 		UserLat:                        userLat,
 		UserLng:                        userLng,
 		TotalPrice:                     int64(totalPrice),
-		EstimatedDeliveryTimeInMinutes: int32(estimatedTime),
+		EstimatedDeliveryTimeInMinutes: estimatedTime,
 	})
 	if err != nil {
 		return result, fmt.Errorf("failed to save estimate: %w", err)
@@ -101,7 +101,7 @@ func (r *PurchaseRepository) CreateEstimateWithOrders(ctx context.Context, userL
 			err = txQueries.CreateEstimateOrderItem(ctx, database.CreateEstimateOrderItemParams{
 				EstimateOrderID: estimateOrderId,
 				ItemID:          parsedItemID,
-				Quantity:        int32(item.Quantity),
+				Quantity:        item.Quantity,
 			})
 			if err != nil {
 				return result, fmt.Errorf("failed to save estimate order item: %w", err)
@@ -117,7 +117,7 @@ func (r *PurchaseRepository) CreateEstimateWithOrders(ctx context.Context, userL
 
 	result.ID = estimate.ID
 	result.TotalPrice = estimate.TotalPrice
-	result.EstimatedDeliveryTimeInMinutes = estimate.EstimatedDeliveryTimeInMinutes
+	result.EstimatedDeliveryTimeInMinutes = int32(estimate.EstimatedDeliveryTimeInMinutes)
 
 	return result, nil
 }
@@ -165,7 +165,7 @@ func (r *PurchaseRepository) CreateOrderFromEstimate(ctx context.Context, estima
 		// All items for the same merchant should belong to the same order merchant record
 		// Use the is_starting_point value from the first item for this merchant
 		firstDetail := details[0]
-		
+
 		orderMerchantID, err := txQueries.CreateOrderMerchant(ctx, database.CreateOrderMerchantParams{
 			OrderID:         order.ID,
 			MerchantID:      firstDetail.MerchantID,
@@ -196,7 +196,7 @@ func (r *PurchaseRepository) CreateOrderFromEstimate(ctx context.Context, estima
 
 	result.ID = order.ID
 	result.TotalPrice = order.TotalPrice
-	result.EstimatedDeliveryTimeInMinutes = order.EstimatedDeliveryTimeInMinutes
+	result.EstimatedDeliveryTimeInMinutes = int32(order.EstimatedDeliveryTimeInMinutes)
 
 	return result, nil
 }
